Fix rename ui directories and update imports
