+++ src/components/ui/Form/EntityEditor.tsx
@@ -4,7 +4,7 @@ import type { FieldKey, FormMode } from "@entities/core/hooks";
 import ReadOnlyView from "./ReadOnlyView";
 import EditField from "./EditField";
 import EntityForm from "./EntityForm";
-import { DeleteButton } from "@components/ui/button";
+import { DeleteButton } from "@components/ui/Button";
 
 type EntityEditorProps<T extends Record<string, unknown>> = {
     /** Titre de la section */
