+++ src/components/ui/Form/EntityForm.tsx
@@ -1,7 +1,7 @@
 // src/components/ui/form/EntityForm.tsx
 "use client";
 import React, { type FormEvent } from "react";
-import { SaveButton, AddButton, CancelButton } from "@components/ui/button";
+import { SaveButton, AddButton, CancelButton } from "@components/ui/Button";
 import { type FieldKey } from "@entities/core/hooks";
 
 type Props<T extends Record<string, unknown>> = {
